requirements.txt
asgiref==3.6.0
Django==4.1.4
django-phone-field==1.8.1
sqlparse==0.4.3
tzdata==2022.7
